requirements.txt
azure-functions
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
urllib3==2.5.0
Werkzeug==3.1.3
requests
